refactor(api): split file transfer handlers
